♻️ [Refactor] Movement 관련 View 파일 분리 및 폴더 분리
